feat: add delete post method

+++ pkg/uploader/uploader.go
@@ -2,8 +2,10 @@ package uploader
 
 import (
 	"context"
+	"fmt"
 	"io"
 	"log"
+	"strings"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/config"
@@ -18,9 +20,10 @@ type Uploader interface {
 }
 
 type uploader struct {
-	manager *manager.Uploader
-	client  *s3.Client
-	bucket  string
+	manager  *manager.Uploader
+	client   *s3.Client
+	bucket   string
+	endpoint string
 }
 
 func MustNew(AccessKey, SecretKey, Region, Endpoint, Bucket string) Uploader {
@@ -36,7 +39,11 @@ func MustNew(AccessKey, SecretKey, Region, Endpoint, Bucket string) Uploader {
 	client := s3.NewFromConfig(config)
 	manager := manager.NewUploader(client)
 
-	return &uploader{manager, client, Bucket}
+	endpoint, found := strings.CutPrefix(Endpoint, "https://")
+	if !found {
+		log.Fatalf("failed to parse endpoint: %s", Endpoint)
+	}
+	return &uploader{manager, client, Bucket, endpoint}
 }
 
 func (u *uploader) Upload(ctx context.Context, key string, body io.Reader) (string, error) {
@@ -49,7 +56,11 @@ func (u *uploader) Upload(ctx context.Context, key string, body io.Reader) (stri
 	return result.Location, err
 }
 
-func (u *uploader) Delete(ctx context.Context, key string) error {
+func (u *uploader) Delete(ctx context.Context, url string) error {
+	key, found := strings.CutPrefix(url, fmt.Sprintf("https://%s.%s", u.bucket, u.endpoint))
+	if !found {
+		return fmt.Errorf("failed to parse key: %s", url)
+	}
 	_, err := u.client.DeleteObject(ctx, &s3.DeleteObjectInput{
 		Bucket: aws.String(u.bucket),
 		Key:    aws.String(key),
